requirements.txt
redis~=7.1.0
requests~=2.32.5
environs~=14.5.0
python-telegram-bot==13.15
standard-imghdr==3.13.0
urllib3==1.26.15
